Add initial support for AST27x0
- start.S and bootrom.ld
The SRAM size is 128KB, with a start address of 0x10000000.
Create a global variable "next_boot_addr" to store the BL31 load address
(Trusted Firmware-A).
When CPU0 calls the "load_boot_image" function, the function will return the
BL31 address, which will be placed in the X0 register. This address is then
assigned to "next_boot_addr" as the BL31 address, and an event is sent to notify
CPU1, CPU2, and CPU3. Afterward, CPU0 jumps to the BL31 address.
Meanwhile, CPU1, CPU2, and CPU3 continuously check if "next_boot_addr" is
non-zero. Once it is set, they also jump to the BL31 address.

- lib/libfdt
To parse the U-Boot FIT image, use the library from the DTC repository:
https://github.com/dgibson/dtc.

- lib/libc/minimal
To build "libfdt/fdt.c" and "libfdt/fdt_ro.c", support for API functions like
"strlen" and "memcpy" is required. These functions are provided by "string.c",
which can be found in the Zephyr repository:
https://github.com/zephyrproject-rtos/zephyr/tree/main/lib/libc/minimal.

- uart_aspeed.c
To develop vBootROM, add this driver to support the serial console.
This allows users to view logs from the console for debugging purposes.
This file is based on the driver located at:
https://github.com/AspeedTech-BMC/zephyr/blob/aspeed-main-v3.7.0/drivers/serial/uart_aspeed.c.
It only support TX.

- uart_console.c`
I created a "uprintf" function, which is very similar to the "printf" function,
for development purposes to display UART logs on the console.

- image.c
This is the main function of vBootROM for AST27x0. It initializes the UART and
registers the console callback function. Next, it searches for the U-Boot FIT
image within the address range of 0x100000000 to 0x100400000.
If the U-Boot FIT image is found, it parses all sub-images in the "/images" node
to retrieve their data and load addresses. The data is then copied to their
respective load addresses. Finally, it returns the BL31 address as the next boot
jump address.

Signed-off-by: Jamin Lin <jamin_lin@aspeedtech.com>